Move the existing services into setup folder
